Move taxi game into crazy_robotaxi package
