+++ integrations/omnidreams/omnidreams/interactive_drive/crazy_robotaxi/__init__.py
@@ -1,0 +1,4 @@
+# SPDX-License-Identifier: Apache-2.0
+# SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+
+"""Crazy Robotaxi gameplay, driving, persistence, and physics policies."""
